requirements.txt
fastapi==0.104.1
uvicorn[standard]>=0.35.0
websockets>=15.0.1
wsproto>=1.2.0
python-multipart==0.0.6
python-dotenv>=1.0.0
google-cloud-speech==2.21.0 